moved also `Communication`-base class to new `communication_backends`-module
