+++ heat/communication_backends/__init__.py
@@ -2,4 +2,5 @@
 Add the communication_backends functions to the ht.communication namespace
 """
 
+from .communication import *
 from .mpi4py4torch import *
