+++ heat/core/__init__.py
@@ -4,7 +4,7 @@ add the core heat function to the ht.core namespace
 
 from .arithmetics import *
 from .base import *
-from .communication import *
+from ..communication_backends.communication import *
 from .constants import *
 from .complex_math import *
 from .devices import *
