+++ heat/core/indexing.py
@@ -5,7 +5,7 @@ Functions relating to indices of items within DNDarrays, i.e. `where()`
 import torch
 from typing import List, Dict, Any, TypeVar, Union, Tuple, Sequence
 
-from .communication import MPI
+from ..communication_backends.communication import MPI
 from .dndarray import DNDarray
 from . import sanitation
 from . import types
