+++ heat/core/linalg/qr.py
@@ -5,7 +5,7 @@ import collections
 import torch
 from typing import Type, Callable, Dict, Any, TypeVar, Union, Tuple
 
-from ..communication import MPICommunication
+from ...communication_backends.communication import MPICommunication
 from ..types import datatype
 from ..tiling import SquareDiagTiles
 from ..dndarray import DNDarray
