+++ heat/core/sanitation.py
@@ -8,7 +8,7 @@ import torch
 import warnings
 from typing import Any, Union, Sequence, List, Tuple
 
-from .communication import MPI, Communication
+from ..communication_backends.communication import MPI, Communication
 from .dndarray import DNDarray
 
 from . import factories
